requirements.txt
beautifulsoup4==4.3.2
Django==1.6.1
django-imagekit==3.2
django-nose==1.2
django-tinymce==1.5.2
django-toolbelt==0.0.1
django-tagging==0.3.2
django-blog-zinnia==0.14.1
django-mptt==0.6.1
django-xmlrpc==0.1.5
gunicorn
mandrill==1.0.53
mock==1.0.1
nose==1.3.0
Pillow==2.3.0
pyparsing==2.0.2
pytz==2014.4
python-memcached
requests==2.2.1
sorl-thumbnail==11.12
South==0.8.4
supervisor
